pyproject.toml
[project]
name = "apantli"
version = "0.3.7"
description = "Lightweight local LLM proxy with SQLite cost tracking & interactive model comparison. Unified OpenAI-compatible API for multiple providers."

license = {text = "Apache-2.0"}
authors = [
    {name = "Philip Borenstein", email = "pborenstein@gmail.com"}
]
requires-python = ">=3.13"
keywords = ["llm", "proxy", "litellm", "openai", "anthropic", "cost-tracking", "sqlite"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: Proxy Servers",
]
dependencies = [
    "fastapi",
    "uvicorn",
    "litellm",
    "pyyaml",
    "python-dotenv",
    "netifaces",
    "tenacity",
    "aiosqlite",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "mypy",
    "types-PyYAML",
    "types-netifaces",
]

[project.urls]
Homepage = "https://github.com/pborenstein/apantli"
Repository = "https://github.com/pborenstein/apantli"
Issues = "https://github.com/pborenstein/apantli/issues"

[project.scripts]
apantli = "apantli.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.mypy]
python_version = "3.13"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

# Start strict, relax as needed
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false